fix: delete useless code
