--- react-components/src/components/form/userInfoCard/types.ts
@@ -1,8 +1,0 @@
-export type CardInfo = {
-  name: string;
-  birthdayDate: string;
-  gender: string;
-  dessert: string;
-  additives: string[];
-  file: string;
-};
