--- react-components/src/components/modal/types.ts
@@ -1,20 +1,0 @@
-export type FullCardProps = {
-  id: number;
-  created: string;
-  episode?: string[];
-  gender: string;
-  image: string;
-  location: {
-    name: string;
-    url?: string;
-  };
-  name: string;
-  origin?: {
-    name: string;
-    url: string;
-  };
-  species: string;
-  status: string;
-  type: string;
-  url?: string;
-};
