--- react-components/src/components/shortCard/types.ts
@@ -1,5 +1,0 @@
-export type ShortCardProps = {
-  name: string;
-  id: number;
-  image: string;
-};
