--- react-components/src/data.ts
@@ -1,157 +1,0 @@
-const cardsData = [
-  {
-    id: 'cake',
-    header: 'Cake',
-    src: '/cake.jpg',
-    description:
-      'With its outstanding vanilla flavor, pillowy soft crumb, and creamy vanilla buttercream, this is truly the best vanilla cake you’ve ever had',
-    rating: 4.5,
-    price: 23.4,
-    popular: true,
-    discount: 5,
-  },
-  {
-    id: 'chocolateCandies',
-    header: 'Chocolate candies',
-    src: '/candy.jpg',
-    description:
-      'Our specialty chocolates keep our customers coming back, with gourmet fillings enrobed in rich, creamy chocolate with a satisfying snap and smooth finish',
-    rating: 4.7,
-    price: 25,
-    popular: true,
-    discount: 0,
-  },
-  {
-    id: 'cookies',
-    header: 'Vanilla cookies',
-    src: '/cookie.jpg',
-    description:
-      'Melt-in-the-mouth cookies with a lovely hint of vanilla. They have a delicate vanilla flavour, which is the perfect accompaniment to a slick layer of icing.',
-    rating: 4.3,
-    price: 12,
-    popular: false,
-    discount: 7,
-  },
-  {
-    id: 'donuts',
-    header: 'Donuts',
-    src: '/donuts.jpg',
-    description:
-      'Once you`ve tasted these irresistibly plump and sugary treats from doughnut master Justin Gellatly, there`ll be no going back!',
-    rating: 4.9,
-    price: 20.5,
-    popular: true,
-    discount: 0,
-  },
-  {
-    id: 'gum',
-    header: 'Gum',
-    src: '/gum.jpg',
-    description:
-      'You may be able to blow small bubbles, but it will be a little different in texture than what your kids are used to in regular chewing gum.',
-    rating: 4.8,
-    price: 11,
-    popular: false,
-    discount: 5,
-  },
-  {
-    id: 'gummyBears',
-    header: 'Gummy bears',
-    src: '/gummy-bear.jpg',
-    description:
-      'Perfectly sweet, chewy, and colorful healthy gummy bears made with real fruit. These Jello gummy bears are such a convincing twist on the classic candy! ',
-    rating: 4.6,
-    price: 15.5,
-    popular: false,
-    discount: 10,
-  },
-  {
-    id: 'iceCream',
-    header: 'Ice cream',
-    src: '/ice-cream.jpg',
-    description:
-      'A caramelly combo of butter and brown sugar drives the flavor in this creamy treat. A shot of bourbon gives it a boost, but it would still be a delicious ice cream if you want to skip the booze.',
-    rating: 4.9,
-    price: 9,
-    popular: true,
-    discount: 0,
-  },
-  {
-    id: 'lollipops',
-    header: 'Lollipops',
-    src: '/lollipops.jpg',
-    description:
-      'Enjoy a rainbow of colors and flavors with these individually wrapped lollipops. The curiosity for making special blends of flavors and colors of candy became a passion for providing people with bold tasting lollipops.',
-    rating: 4.5,
-    price: 6,
-    popular: false,
-    discount: 0,
-  },
-  {
-    id: 'macaron',
-    header: 'Macaron',
-    src: '/macarons.jpg',
-    description:
-      'Colorful sprinkles, pretty in pink, and a hint of nostalgia... We`ve taken those classic animal crackers you used to have at school lunch and gave them a fun and tasty new twist!',
-    rating: 5,
-    price: 2.5,
-    popular: true,
-    discount: 0,
-  },
-  {
-    id: 'marshmallow',
-    header: 'Marshmallow',
-    src: '/marshmallow.jpg',
-    description:
-      'Talk about a party in a box! Our Funfetti marshmallows taste just like birthday cake, so you can celebrate anytime! We believe every day is the perfect day to Live Your Sweet Life.',
-    rating: 4.8,
-    price: 12.5,
-    popular: false,
-    discount: 0,
-  },
-  {
-    id: 'meringue',
-    header: 'Meringue',
-    src: '/meringue.jpg',
-    description:
-      'Our meringue kisses are crispy on the outside, mallowy in the middle and made with all natural colours and flavours. Natural colours and flavourings used – free range egg white, caster sugar, natural colourings, and flavourings.',
-    rating: 4.9,
-    price: 15.6,
-    popular: false,
-    discount: 5,
-  },
-  {
-    id: 'muffins',
-    header: 'Muffins',
-    src: '/muffins.jpg',
-    description:
-      'Breakfast with freshly baked healthy muffins and an espresso for a daily delectable treat. Packed with healthy ingredients like fruits, nuts and available in a variety of flavors.',
-    rating: 4.8,
-    price: 9.4,
-    popular: false,
-    discount: 5,
-  },
-  {
-    id: 'raspberry',
-    header: 'Raspberry',
-    src: '/raspberry.jpg',
-    description: 'Just fresh sweet raspberry. A great raspberry in your life!',
-    rating: 5,
-    price: 8,
-    popular: true,
-    discount: 0,
-  },
-  {
-    id: 'strawberryInChocolate',
-    header: 'Strawberry in chocolate',
-    src: '/strawberry-in-chocolate.jpg',
-    description:
-      ' All our berries are hand dipped only in real premium chocolate and beautifully packed in our signature gift box with ice packs, then shipped overnight to ensure perfect arrival.',
-    rating: 4.9,
-    price: 15,
-    popular: false,
-    discount: 0,
-  },
-];
-
-export default cardsData;
