+++ react-components/src/store/collectModalDataSlice.ts
@@ -1,5 +1,25 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { FullCardProps } from '../components/modal/types';
+
+export type FullCardProps = {
+  id: number;
+  created: string;
+  episode?: string[];
+  gender: string;
+  image: string;
+  location: {
+    name: string;
+    url?: string;
+  };
+  name: string;
+  origin?: {
+    name: string;
+    url: string;
+  };
+  species: string;
+  status: string;
+  type: string;
+  url?: string;
+};
 
 const defaultModalData = {
   id: 1,
